gcloud-native/cloud-run/src/requirements.txt
google-cloud-storage==2.7.0
slack-bolt==1.16.2
slack-sdk==3.20.0
uvicorn==0.20.0
fastapi==0.92.0
